invoice_extraction_tool/requirements.txt
streamlit==1.45.0
pandas==2.2.3
pytesseract==0.3.13
pdf2image==1.17.0
pillow==11.2.1
pdfplumber==0.11.6
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
numpy==2.2.5